github-actions: fix nightly version-checking
A nightly package usually looks like this:
`3.35.1.449.ge27fb36-snapshot+20220301T230257`

However we forgot that the git describe part of the version (`.449.ge27fb36`) can be empty
if the nightly package is built on a recent tag.
Now the version for the nightly package is this:
`3.36.1-snapshot+20220306T230226`

In short, if the nightly package is created on a tag, we shouldn't
expect the git-describe part of the version, so I've made that part of
the regex optional.

Signed-off-by: Gabor Nagy <gabor.nagy@oneidentity.com>