test-apt-packages: allow smaller number in version-checking's commit distance part
I didn't check before that the package version check had a very strict quantifier check
in the git describe part of the version (e.g. 3.35.1.361):
it required that the commit distance to be 3 digits long.

I've relaxed the criteria for that and also where applicable (e.g. patch
version only allowed to be one digit long too).

Some examples of the recent package versions:
3.35.1.361.g1cd66f0-snapshot+20220226T230235
3.36.1-snapshot+20220304T230252
3.36.1.2.g134cd38-snapshot+20220307T230230

Signed-off-by: Gabor Nagy <gabor.nagy@oneidentity.com>