logmsg: unset stale match values
If we set a match that is outside of the currently tracked match range,
unset elements that are between the old and the new bounds.

E.g. if we have $1, $2, $3 set and then set $5, the value of $4 would be
undefined and would hold a value that leaks out from a previous operation.

Make sure that $4 is unset in this case.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>