cfg-tree: fix persist-name uniqueness validation
The hash used for validating the persist names stored the original pointer
as returned by the log_pipe_get_persist_name() function, which is usually
a static string. This means that if we have two instances of the same
object in our config, its persist name will be formatted in the same spot,
effectively changing the key as stored in the GHashTable.

Clearly this is not what we want. Make sure we strdup the key before storing
it.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>