afstomp: always use "0" in the receipt header
After 3d124fb the uniqueness of the `seq_num` field is not guaranteed.
In theory this could cause a problem in the STOMP communication, but
with the current implementation it works correctly.

With this commit I intend to emphasize that we do not need a unique ID
in the receipt header by using the same one each time. I also added a
comment which explains why it is not a problem currently, and how it
could cause a problem in the future.

------------------------------------------------------------------------

Explanation:

We check the server's response before sending a new frame to it.
Because of this, we do not need a unique receipt header.

This changes if multiple workers and/or batching support is introduced.
Make sure to use a unique receipt-id if one of the above gets
implemented.

------------------------------------------------------------------------

Useful links:

https://stomp.github.io/stomp-specification-1.2.html#Header_receipt
https://stomp.github.io/stomp-specification-1.2.html#RECEIPT

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>