dbld: fix missing repositories in workaround_rpm_repos
Some repositories (updates, updates-modular and fedora-modular) were
using the same URL as the "fedora" repository, which caused that some
package updates were not visible, leading to package conflicts with the
existing package installation.
You can check with `dnf repolist --verbose` to see that current mirror
(if no baseurl is used in the .repo file), or query the metalink URL.

Also, we should use primarily the metalink service instead of disabling
it, and only fallback to using a specific mirror.

Signed-off-by: Gabor Nagy <gabor.nagy@oneidentity.com>