groupingby: remove state sharing between clones
After careful consideration, we decided not to share the state between
grouping-by() clones, which get created if the same parser statement is
referenced from multiple parts of the configuration. The rationale is
summarized below.

Example configuration:

@version: 3.36

parser p_gby {
        grouping-by(
                key("$PROGRAM")
                aggregate(
                        value("aggr" "$(list-slice :-1 $(context-values $mod))")
                )
                timeout(3)
        );
};

log {
        source { tcp(port(2000)); };
        if ("$(% $SEC 2)" == "0") {
                rewrite { set("0" value("mod")); };
                parser(p_gby);
        } else {
                rewrite { set("1" value("mod")); };
                parser(p_gby);
        };
        filter { "$aggr" ne "" };
        destination { file("aggr.log" template("$(format-json aggr)\n")); };
};

With the example config above, parser object named "p_gby" is referenced
twice, in two parallel branches of the configuration:
  1) once when the timestamp has an even second value,
  2) once when the timestamp has an odd second value

If state sharing would be implemented, both references would store messages
in the same state object, e.g. aggregation contexts would contain
a mixture of messages, received using either of the two branches. This
consequence is already pretty difficult to follow and can be confusing.

To add to the confusion, another problem comes up. Once the timeout in
grouping-by() elapses, which branch would emit the "aggregate" message?
Following our existing configuration logic, we should emit the aggregate
twice. This means that our destination in the example config above, would
receive the same, identical message twice both of which contains a mixture
of messages.

With that in perspective we decided not to share the state:
  * this matches other existing cases, db-parser() or rate-limit() does
    not share state either.
  * with unshared state, we'd basically have two grouping-by() instances
    that are configured the same but with distinct states.
  * each grouping-by() clone would process messages only on its
    respective branch
  * each aggregate would contain a distinct set of messages, they are
    not mixed up

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>