github-actions: do not try to automatically push docker images
Currently, we try to push an image even if the build step failed.

This is a leftover of when the image build job was refactored in
[edb1d06](https://github.com/syslog-ng/syslog-ng/pull/3932/commits/edb1d06371d1a3ed15f3ceacf1a79ae86de88dfa).
Previously we built the images in 1 job concurrently
and then tried to push the built images.
Since even 1 image build resulted in a job failure,
we used the `always()` condition in the later steps.

Signed-off-by: Gabor Nagy <gabor.nagy@oneidentity.com>