groupingby: refactor _perform_groupby() to not return "success"
The decision on whether an incoming message is to be dropped should not
depend on "group bying" as it always consumes a message. It is however
desirable to make this decision one level up, in grouping_by_process(),
so let it do that.

An upcoming change will allow the dropping of messages that make up
a correlation context.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>