libtest/mock-logpipe: extract from test_grouping_by
This also involved a license change, as this code was originally embedded
in a GPL area (modules/dbparser/tests) and now it gets moved to libtest,
which is under the LGPL.

This code has originally been introduced by myself in a previous PR and
then improved even in this PR, nonetheless it is written by my in its
entirety, meaning that I am the copyright holder. And as such I can relicense
it so it can become part of libtest.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>