logmsg: unset matches proactively
Instead of relying on num_matches to determine if a match has a value or not,
unset unneeded matches proactively whenever we are dropping the previous
set of matches.

This in slightly in contrast with the preceeding patches in the same branch,
where I got rid off proactive cleaning in favour of using num_matches.

The reason I've decided against the first implementation was that there
are multiple code-paths that give access to name-value pairs (that include
matches).

 1) log_msg_get_value() and related functions
 2) log_msg_get_match() and related functions
 3) log_msg_values_foreach()

I did solve all of these cases, however I felt that handling that handling
these cases in all of the above was

 1) complicating things
 2) pushed extra work to the fast-path (instead of the slow path).

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>