templates: add log_template_get_trivial_value_handle()
This patch adds a new function to LogTemplate to be able to query the
name-value handle behind a "trivial" template.  A trivial template is one
which only references a single name-value pair.

This is useful when we want to use the value of that name-value pair without
having to format the template, that'd result in a string allocation and
copying of the data.

The patch refactors log_template_get_trivial_value{_and_type}() so that we avoid
redundancy between these two functions.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>