msg-format: resolve feature envy by moving log_msg_new() to msg_format_parse()
log_msg_new() was actually a feature envy and should have been a method
for MsgFormat: it takes an MsgFormat instance and input data, and
based on this it created and populates a LogMessage instance.

Move this functionality from LogMessage to MsgFormat, straightening
the dependencies in the right direction.

Albeit the change is trivial, it cascades into a lot of files, primarily
into test code.

Test code should avoid doing syslog parsing anyway but in a lot of
cases those still use it.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>