Correct logrotate not reloading syslog-ng
The existing logritate config uses a postrotate command of:
  invoke-rc.d syslog-ng reload > /dev/null

This fails for several reasons.

A working equivalent is:
  syslog-ng-ctl reload

This commit simply replaces the existing /etc/logrotate.d/syslog-ng file
with an exact copy that has the failing command swapped out for the
working one.

The consequence of the failing command is that if cron is started,
logrotate will be called and will rotate the /var/logs/messages log
written by syslog-ng but syslog-ng will continue writing to the old log
because it fails to be reloaded.