packaging/debian: Correct logrotate not reloading syslog-ng
The existing logritate config uses a postrotate command of:
  invoke-rc.d syslog-ng reload > /dev/null

This fails for several reasons.

A working equivalent is:
  syslog-ng-ctl reload > /dev/null

This commit simply swaps out the failing command for a working one

The consequence of the failing command is that if cron is started,
logrotate will be called and will rotate the /var/logs/messages log
written by syslog-ng but syslog-ng will continue writing to the old log
because it fails to be reloaded.

Signed-off-by: Ben Burrows <ben@benburrows.com>