github-actions: use Private Access Token in draft-release
Anything that is triggered by the Github Actions provided GITHUB_TOKEN
cannot trigger other workflows. We have seen this happening with PRs
opened by a workflow, but it seems like it applies for releases, too.

In my experience, when the draft-release workflow running with
GITHUB_TOKEN creates a draft release, it does not matter that a human
publishes that release manually, the stable-release workflow will not be
triggered.

Signed-off-by: Attila Szakacs <attila.szakacs@oneidentity.com>