modules, logwriter: extract reopening of kept-alive connections
log_writer_init() was asymmetrical: in case the LogWriter's proto
were kept-alive during reload
(e.g. a connection in afsocket, or a program in afprog),
the LogWriter class reopened __itself__ with it's stored proto at init phase.

I've extracted this to remove the hidden complexity,
so the owner of LogWriter can reopen it after the required LogProto is acquired,
and the AFSocketDestDriver's state is restored as well.

My understanding is that LogWriter should be controlled by its owner,
thus we can prevent that a LogWriter is running while its owner is
half-initialized.

Signed-off-by: Gabor Nagy <gabor.nagy@oneidentity.com>