afsocket: store dest_addr in case of reload
This also fixes a crash in case of kept-alive connections across reloads,
when the remote host closes the connection, and previously
the name resolution of the destination address failed.

AFInetDestDriver has to check for an already opened connection
during it's init phase (afinet_dd_setup_addresses).
The destination driver's initialization can't be skipped
in case pf a re-used connection, we can only skip steps
that is related to the re-used connection (e.g. it's address).

Signed-off-by: Gabor Nagy <gabor.nagy@oneidentity.com>