afsocket: move connection initialized flag setting
Some refactors were made in afsocket since `connection_initialized` were
introduced, and I think it's bad that it remained in
`afsocket_dd_setup_connection()`.

Origin commit of `connection_initialized is 006b3d2ac0 from 2015.
Looking at how afsocket_dd_setup_connection()` looked like, it seems we set `connection_initialized` when we have a working LogWriter,
and doesn't mean it's an established connection.
(working means it's initialized hence has a queue, but no need to be opened)

`connection_initialized` is a state that is used to check if we have a
working LogWriter that needs to be kept across reloads.

Signed-off-by: Gabor Nagy <gabor.nagy@oneidentity.com>