all: Refactor evt_tag_printf() to evt_tag_mem() where possible
In many places, evt_tag_printf() was only used to print a string/buffer
with limited length (instead of null-terminated string). This can be
refactored to evt_tag_mem(), which does exactly that - print a limited
buffer as ASCII string.

There are some uses remaining which can not be trivially converted,
for example evt_tag_printf(..., "%.*s%s", ...).

Signed-off-by: Daniel Klauer <daniel.klauer@gin.de>