java: remove native check for named templates
Named templates are not supported in Java destinations.
A native check is meant to notify users about this, but this is a
cross-module implementation, which would be hard to fix
(the native method is implemented in "java", but the class itself is
located in "java-modules"), so the check has been removed and a more
general error message will be displayed instead.

Signed-off-by: László Várady <laszlo.varady@protonmail.com>