cfg-grammar: add log-level() global option
This patch adds a feature to control the internal verbosity of syslog-ng
(e.g. verbose, debug, trace) via the configuration file.

Example:

options {
  log-level(debug);
};

Possible values for log-level:
  - default, just normal log messages
  - verbose, normal + messages that get enabled via -v
  - debug, verbose + messages that get enabled via -d
  - trace, debug + messages that get enabled via -t

The command line continues to dominate the log-level, e.g.  if the command
line contains options to control the level of logging, the configuration file would be ignored.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>