generic-number: make sure all operations on NaN trigger an exception
The calling code should check if the GenericNumber is a NaN before calling
gn_is_zero() or gn_compare(). Although gn_is_zero() could be evaluated
(a NaN is never zero), there's no appropriate return value for
gn_compare() that could indicate that this is not a valid operation.

Make all of these fail with a SIGABRT.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>