filter-cmp: change the truth table when comparing against null
null values are only converted to a numeric value (e.g. 0) if it's compared
with less than or greater than operators. When checking for equality
or non-equality, we check if the target value is null too.

Equality and non-equality:

null == null is true
null != null is false
<non-null> == null is false
<non-null> != null is true

Numeric comparisons (null converts to 0 in these cases):

null < 1 is true
null > -1 is true

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>