github-actions: calculate tag for nightly image
`stable` tags will work like it did before.
`nightly` tag will be `nightly`. :)

It would be nice to have custom tags for each nightly build, but it
would require some clean-up logic, which is not trivial. So we decided
to have this for now, and do the advanced tagging later.

Signed-off-by: Attila Szakacs <szakacs.attila96@gmail.com>