csv-parser: add parsing to log message matches
Instead of always having to come up with names for the columns we parse,
this patch makes it possible to parse into the "matches" array, e.g.
$1, $2, $3...

To do this, simply omit the previously mandatory columns() option to
csv-parser().

	@version: current

	log {
		source { tcp(port(2000) flags(no-parse)); };

		parser { csv-parser(delimiters(',')  dialect(escape-backslash)); };
		destination { stdout(template("$ISODATE $*\n")); };
	};

This makes the parsed content also available as a syslog-ng list, as
the "$*" macro can be used to produce a list out of the parsed values.

Signed-off-by: yashmathne <yash6866@gmail.com>
Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>