mainloop-control: translate legacy LOG commands invoked via syslog-ng-ctl to log levels
syslog-ng-ctl is issuing separate commands for various log levels, e.g.

LOG VERBOSE ON
LOG DEBUG ON
LOG TRACE ON

which was operating on various verbosity flags exported by the messages
module.  This patch changes the implementation so while the interface
continues to work, we are mapping the control commands to log levels.

Whevever we turn ON verbose/debug/trace, we set the log level to that
specific values.  When we turn them OFF we set the log level one less than
the specific value.

LOG VERBOSE ON => verbose
LOG VERBOSE OFF => default
LOG DEBUG ON => debug
LOG DEBUG OFF => verbose
LOG TRACE ON => trace
LOG TRACE OFF => debug

This keeps compatibility with old syslog-ng-ctl binaries with a tiny change
in functionality: the user will not be able to enable trace messages while
not enabling debug at the same time.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>