syslog-ng-debun: Drop direct access to dpkg internal database
Files in /var/lib/dpkg/info/ is dpkg internal files. It can create some
issues accessing it directly, but the most important one is that the dpkg
developers would like to change the structure.

For this reason, it is better to validate the integrity of the installed
syslog-ng packages using the official interface (dpkg --verify).

This patch resolves #4139

Signed-off-by: Attila Szalay <sasa@debian.org>