python: instead using a PersistState related PyCapsule, store a GlobalConfig reference
Each syslog-ng config has a separate "view" of the Python execution
context, namely a separate main module. Whenever we reload syslog-ng,
we will reconstruct its main module as well.

This means there's a one-to-one relationship between
GlobalConfig <> Python context.

In some cases, we would need a reference to the actual config from
the Python side, one such case is the PersistState wrapper.

This patch creates a more generic way to access the GlobalConfig instance
instead of the PeristState specific solution we currently use.

The new system exports the current config as _syslogng_main.__config__.
This variable is a PyCapsule (similarly to _syslogng.persist_state) but
pointing to GlobalConfig.

With that we establish a circular reference:

GlobalConfig -> PythonConfig -> main_module -> main_module.__config__ -> GlobalConfig

This does not cause leaks as GlobalConfig is not reference counted, its
lifetime is controlled by the reload machinery. But as long as we execute
Python code, we are always bound to a GlobalConfig instance, ensuring its
existence.

The patch also removes the need to update _syslogng.persist_state
at init time.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>