python: clone LogTemplateOptions into PyLogTemplateOptions
LogTemplateOptions can be specified to a LogTemplate() instance
at constructor time and this reference is then retained until the template
exists, so we need to ensure that LogTemplateOptions remains valid.

LogTemplateOptions however is not reference counted and in a lot of cases
is a member of an encapsulating, larger structure.

To avoid complications, this patch changes PyLogTemplateOptions so that the
underlying C struct is cloned into a local copy.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>