news: remove fixed contributor list from newsfile generator script
It was a naive idea of mine a couple of years ago to maintain a fixed
contributor list as a sign of our appreciation for the members of the
syslog-ng development team.

As it turned out, nothing is really "fixed", nor was the list.

Removing the hard-coded list, so we don't have to update it all the
time.

So long.

Signed-off-by: László Várady <laszlo.varady@protonmail.com>