dbld/builddeps: use --ignore-installed option when updating pip/setuptools
On centos-7 the pip installation of setuptools removed the rpm installed
setuptools package from /usr/lib. The problem with this is that under
rpmbuild the syslog-ng build scripts don't have /usr/local/lib/pythonX.Y
in their PYTHONPATH, causing the build to fail, as there's only a broken
setuptools package under /usr/lib.

This option is also passed to the other invocations of pip as well and
fixes the issue, as pip will not attempt to remove the old package.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>