python: remove unit test specific _py_init_interpreter() functions
This is just copy-paste code that we could reuse from the plugin, use that.
It is true that _py_init_interpreter() will initialize more things than
this specific implementations, however I am about to complicate the
one in the plugin, which I wouldn't like to replicate in the unit tests.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>