python: add support for initializing a production virtualenv
The original idea has been to automatically populate our virtualenv at
syslog-ng startup, however I had to realize that it would execute third
party installation code as root on end-user systems.  Not good.

The solution implemented in this patch pushes the responsibility of Python
depedencencies to the user (or the package maintainer), while giving a
simple to use tool to automate the creation/maintenance of a Python
virtualenv which syslog-ng Python packages can then use.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>