tests/copyright/check.sh: move whether the file is text to after the list was pruned
There's a "grep" call in the find command line to limit the set of
results to text files (as grep defines it). The problem with this is
that I have a huge number of files in dbld/ subdirectories, which are
all read by grep to determine if they are text files.

Move the grep call inside the loop and add the dbld subdirectories into the
IGNORE list. This makes the script usable on my source tree
(just barely though)

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>