logproto-proxied-text-server: refactor fetch_into_buffer() to use a state machine
Since I am about to add support for the PROXYv2 protocol, I need to be
a bit more flexible how we fetch the proxy header. Refactor the fetch
function to be a state machine, which can later be extended.

There's one change functionality wise: one previously accepted case, where
we only had a CR character without an NL is now rejected, instead
of accepted. Earlier, the only reason it worked was that the
unit test contained an EOF right after the CR character, causing the
proxy header processing logic to process the PROXY line, even though there
was no NL character.

The current change however would return LPS_EOF as we receive it, so
the PROXY line is not processed.

This patch contains a change in the unit test to reflect the new reality.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>