configure.ac: add --with-python-packages=none
This change adds a new setting for --with-python-packages: "none" specifies
that we should not install dependencies in any way while installing our own
modules.  This allows the downstream distributor to include our Python
modules without having to deploy their Python dependencies in the build
environment.  This will mean that unit tests will not be able to execute,
but apart from that the code is installed.

This can then be further customized by culling code that does not apply to a
specific target platform.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>