mainloop-call: extract main_loop_wait_for_pending_call_to_finish()
As a preparation for fixing #4258, this patch extracts a few lines as
a new function from main_loop_call().

The goal of the new function is to wait for any pending executions of
mainloop-calls, which can happen if a worker thread submits a new one while
the old is still executing. At the moment, there can be only ONE outstanding
mainloop-calls at the same time.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>