mainloop-call: fix crash by waiting for pending mainloop-calls at thread deinit time
As explained in #4258, a crash can happen in syslog-ng if a worker thread
exits while a mainloop call is still being processed.

The solution is to prevent the worker from exiting until that pending call
is finished.

NOTE: in this state the worker is not going to execute tasks, but for all
ivykis knows this thread does not really exist anymore. So if the workpool
requires new threads, they are going to be launched, while this exists
in limbo.

Reported-by: @chucksilvers
Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>