patterndb: drop "synthetic" argument to the emit function
Earlier patterndb called emit() for all messages, either the "originals"
that we ingested into our context and also for "synthetics" e.g. those
we generated based on our correlation rules.

The only reason to do this was the "unit" test for patterndb, which
asserted on messages in a specific order, but the production code never
used this property and the production ordering differed from the one
encoded in the unit test.

With this change, the unit test would just append the incoming message to
the end of the output array, which flips some of the indexes in testcases,
used to reference one specific message.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>