grouping-parser: emit both timed out and synthetic messages _after_ the grouping operation
Due to the sequence of refactoring steps, the collection of emitted messages
was sectioned into two separate collect/flush steps, once for timed out
contexts and then once for synthetic messages.

This can cause problems, especially if the message is being changed during
the grouping operation (which is now the case, due to ${.classifier.context_id}),

What happens that even if we clone the incoming message and make it writable,
it becomes read-only as it hits the destination queues.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>