grouping-parser: avoid making LogMessage writable
We made every message that we consumed into out context writable, because
we set ${.classifier.context_id} as a name-value pair in the message
itself. This functionality is coming from db-parser (which was the original
home for correlation) and that implementation predates the $CONTEXT_ID
macro.

$CONTEXT_ID differs from ${.classifier.context_id} in that it is taking
this information from log_template_format() arguments, instead of from the
message. By using $CONTEXT_ID we can avoid changing the message, which can
be a great deal in performance (as it usually would trigger cloning
of the message).

While it is true that the correlation key can be useful while doing the
aggregation, but providing this information in $CONTEXT_ID should
suffice. The setting of ${.classifier.context_id} was only documented for
db-parser and not for grouping-by().

I haven't found any occurence of its use using google (the only hits were
db-parser related, which retains this functionality).

This change should be added to the NEWS file but I don't expect any actual
breakage and the fix for such breakage should be as simple as changing
the template to use $CONTEXT_ID or using an explicit set().

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>