dbld: install criterion from source on Debian testing
Criterion was dropped from Debian testing because of an FTBFS (failed to
build from source).

Criterion fails to build, as there was a minor change in nanopb command line
processing, which causes the criterion.options file to not be found while
processing the protobuf declarations.

This is the fix against criterion, which if applied, this entire patch can
be reverted.

--- criterion-2.4.1/src/protocol/gen-pb.py 2022-04-26 06:51:30.000000000 +0000
+++ criterion-2.4.1.patched/src/protocol/gen-pb.py 2023-02-04 12:51:27.398255067 +0000
@@ -17,7 +17,7 @@

 cmdline = [protoc_path, '--plugin=protoc-gen-nanopb=' + gen_path,
       '-I' + source_dir,
-      '--nanopb_out=' + '-f ' + options_file + ':' + build_dir,
+      '--nanopb_out=' + '-f' + options_file + ':' + build_dir,
       proto_file]

 sys.stderr.write(" ".join(cmdline) + "\n")

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>