logpipe: mark user-created LogPipe instances with PIF_CONFIG_RELATED
While we compile the config, the compiler may generate various LogPipe
instances to perform routing and to connect the LogPipe elements.

This is a bit frustrating when debugging the configuration, as we get
stopped by "unknown" points in the config and there's no means to examine
it.

A subsequent patch will cause the debugger to ignore any non-user generated
LogPipe instances.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>