stats: create stats() global config option block
We have a significant number of stats related global options.
I am planning to add another one, so I felt like it is time to group
these options together.

This change breaks the behavior of the old stats() global option, which
was an alias for stats-freq(), but it has been deprecated for more than
a decade now, I think it is safe to remove it now.

The old stats-...() options are still supported for compatibility
reasons, but new options will be introduced inside the stats() block.

Signed-off-by: Attila Szakacs <szakacs.attila96@gmail.com>