logmpx: fix matched logic
We can have a multiplexer with both pipe_next and next_hops.

The path_options's matched field must only be set to false if and only if
every next_hop and also the pipe_next failed to deliver. If at least one
next_hop or at least the pipe_next delivered it, it must not be set
to false.

With the old logic, if the multiplexer had a pipe_next, the next_hops'
delivery result was ignored completely, which is not the correct behavior.

Signed-off-by: Attila Szakacs <szakacs.attila96@gmail.com>