logpipe: add mechanism to handle junctions as single filtering elements
On a specific log path, we might be using a junction to classify
a log message along multiple alternatives.


                 / alt1 \
path -> ... ->  +--alt2--+ -> ... -> filter -> ...
                 \ alt3 /


In this case, the entire path is "matched" if none of the filtering
elements drop the message.

In the case of a junction, we should set matched to FALSE, if neither
the alt1, alt2 or alt3 branches match.

What might happen though is that one of our branches actually lets the
message through (for example alt1) but then a subsequent filter drops it.

What would happen in this case is that the junction would retry with alt2,
even though alt1 was already successful and it was not the junction itself
that decided the message did not match.

This will cause alternative paths to be considered even though the problem
is outside of the junction. Also, those alternatives may have side-effects
(e.g. the delivery of the message to a destination).

The solution is to make sure that once the message leaves the junction,
our "matched" pointer should not propagate any further, rather any
further processing should rather change the "matched" of our parent
log path.

The solution is to add "outer_matched" pointer to LogPathOptions and that
the last LogPipe of the junction (the join_pipe) will switch the the outer
value. This means that even though the log path may not match as a whole,
the junction itself will not try alternative branches.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>