cfg-tree: use a dedicated LogExpr for conditionals
Add a dedicated LogExprNode for conditionals, so we can delay building the
exact LogPipe structure until compilation time.

Previously a conditional was nothing but a junction, each branch
being "final", conditions being inserted into the LogPipe sequence.

The new structure is aware whether we are using simple ifs (e.g. if (filter-expr)) or
compound ifs (e.g. if { }). The two needs to be compiled differently.

A compound if is the same as it was, a junction, each branch is final,
filtering happens as the body of the if.

A simple if however is different, once the if (filter-expr) matches, we
should not wonder to alternative branches, even if the body of if does not
match at the end.

Therefore we need a midpoint within the conditional: after the filter and
before the body of the conditional. Once we reach this point, we are
considered "matching", any subsequent filters will only impact the parent
log path.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>