logmpx: add explicit "delivery_propagation" setting
Instead of relying on the value of pipe_next, use an explicit setting
to control delivery propagation to the parent log path.

This should be set in case we are a real junction and as such the multiplexer
as a whole is performing a filtering function.

This should be unset in case the multiplexer just serves to dispatch
messages to a set of destinations (e.g. destination reference or inline
destinations).

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>