qdisk: fix setting file_size after prealloc
Setting file_size must be done after each operation, which can alter the
file's size. Creating the header on the disk is an operation like that,
but we cannot blindly set the file_size to the header's size, because
we might have preallocated the file before that which could have already
set file_size.

This commit fixes this issue by checking if file_size is already set or
not, and only setting it to the header's size, when necessary.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>