logqueue-disk-reliable: move memory queue emptying to save_queue()
These memory queues contain LogMessages which are also on the disk-queue
file. Their position in the disk-queue file is also stored in the memory
queues. This means that the memory queues' lifecycles are the same as
the QDisk's, so when we call qdisk_stop(), we should free those queues.

Before this commit in case of a corrupted disk-queue restart, the memory
queues were not emptied, so they contained invalid LogMessages.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>