unit-tests: add 5 min timeout for all tests which did not have any
I faced another stuck unit test in our CI run, and it was a pain to
determine which one was it, because there is no "test started" message.

I think it would be a good habit to set a timeout for all of our existing
and new unit tests. 5 minutes should be plenty enough, maybe we can even
make it 1 minute. There is no reason to make the CI run for 1 hour if
a unit test gets stuck.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>