stateful-parser: messages consumed into a correlation context should not be considered unmatching
As a followup to #4366 and #4058, another edge case was identified
where syslog-ng evaluation handles matched messages incorrectly.

If we are using a correlation while processing the message, the incoming
messages are consumed into a context and then the individual messages
get dropped, at least when using inject-mode(aggregate-only).

In these cases we should not consider dropping of these messages to
be unmatching, as dropping happens as a part of correlation.

This patch adds a light based testcase to exercise this problem and includes
a fix, so that messages consumed into corrlation in
inject-mode(aggregate-only) mode will not cause those messages to traverse
any alternative paths.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>