logqueue-disk-non-reliable: move memory queue emptying to save_queue()
The messages in these memory queues are written to the end of the
disk-buffer file after qdisk_stop(). This means that the memory queues'
lifecycles are the same as the QDisk's, so when we call qdisk_stop(),
we should empty those queues.

Before this commit in case of a corrupted disk-queue restart, the memory
queues were not emptied, so they contained LogMessages, which were also
stored on the corrupted disk-buffer file. Sending them out would cause
message duplication, when the messages from the corrupted files are
restored and sent.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>