qdisk: do not consume messages in save_state()
Prior to my this PR, the messages were freed in
_save_queue() -> qdisk_save_state() -> _save_queue(), then there were
unnecessary frees in _free() -> _free_queue(), and no one decreased the
memory usage counter.

With the previous commits I have moved the freeing to _stop(), and the
memory usage is decreased there as well, but the qdisk_save_state()
branch frees it before we try to do it in _stop(), so there are no
messages there to free, and we cannot correctly decrease the memory
usage counter.

This commit changes this, so we no longer free the messages in
qdisk_save_state(), so they are only freed in _stop() -> empty_queue(),
where the memory usage counter is decreased.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>