afsocket: call afsocket_sd_start_watches() from sd_open_dgram()
afsocket_sd_start_watches() was only called for STREAM sockets but neither
its name nor the mechanisms behind this call implied this fact. Let's
call this from sd_open_dgram() as well.

At the moment this does nothing, as listen_fd is -1 and
the dynamic_window_pool is NULL, but I am moving the packet stats timer into
this function in a subsequent commit.

Signed-off-by: Balazs Scheidler <balazs.scheidler@axoflow.com>