qdisk: hold lock, while calculating next disk-queue filename
There is a race condition, as two threads or two syslog-ng instances
can run _next_filename() simultaneously, and they can get the same
filename.

To mitigate this I have introduced an flock, which solves the issue
between two syslog-ng instances, but does not solve it for threads in
the same syslog-ng instance (man 2 flock):

> If  a  process  uses  open(2)  (or similar) to obtain more than one file descriptor for the same
> file, these file descriptors are treated independently by flock().  An attempt to lock the  file
> using one of these file descriptors may be denied by a lock that the calling process has already
> placed via another file descriptor.

To solve this, I have introduced a static lock, too.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>