logqueue: register stats counters in ctor
LogQueue currently only has driver level stats counters, the queue level
counters are only atomic counters, but do not have StatsCounterItems
assigned to them.

I am planning to introduce queue level stats counters, which cannot be
done because of the following reasons:
  * The counters of LogQueue's (currenlty only driver level ones) are
    registered in the log_queue_register_stats_counters() function.
    If we registered the owned StatsCounterItems there, we would not be
    able to set their values between the ctor and the
    register_stats_counters() calls. This is not a problem currently,
    because we only have atomic counters for the owned counters, which
    do not have to be registered, but when they become real
    StatsCounterItems, they would not be modifiable. This applies for
    new, implementation (e.g. LogQueueDisk) specific counters, too.
  * The necessary info for the StatsClusterKey of the owned counters are
    only available in different levels on the call chain. For example
    driver_id and driver_instance are available in LogWriter and
    LogThreadedDestDriver, an optional worker_id is available only in
    LogThreadedDestDriver, the names of the counters ("memory_queue" or
    "disk_buffer") are available in LogQueueDisk and LogQueueFifo.

To make our lives easier, I have moved the registration of the counters
to the LogQueue ctor, and made it to use the Builder class. It can be
seen, that no one modifies the builder on the call chain (const), but
the Builder for the owned key will be modifiable.

------------------------------------------------------------------------

Although the registration of the counters are now in the ctor, there is
still a need to initialize those values after creating the queue, and
cleanup the values before freeing the queue. The responsibilities are
the following:
  * The counters are registered in the LogQueue ctor.
  * If a LogQueue implementation does something related to these, they
    must update them themselves. A new FIFO instance is always empty, so
    there is nothing to do there. A new disk-buffer instance might start
    with an existing disk-queue file. In log_queue_disk_start() we
    update the queued counter with get_length(), and the non-reliable
    disk-buffer updates the memory_usage counter during its start().
  * The counters are unregistered in the LogQueue dtor.
  * The LogQueue implementations can reset these values, before they are
    getting destroyed, for example when they are getting released
    (log_queue_disk_stop() resets these), but in the LogQueue dtor we
    reset these automatically.

With these, now the responsibility is not on the call site (LogWriter
and LogThreadedDestDriver) to call the unregister() at the correct time.

------------------------------------------------------------------------

As the commit message suggests, unfortunately this commit does multiple
things at once, but my understanding is that they have to change at the
same time to make sense. To make reviewing a bit easier, at least I
separated the unit test follow ups. Sorry for the huge and complicated
commit.

Signed-off-by: Attila Szakacs <attila.szakacs@axoflow.com>